Add compound index and dynamic part-select edge triggers (#424)
- Introduce IndexPlanOp bytecode for index expression eval
- Replace single-slot LateBoundIndex with plan + dep_slots
- Support @(posedge bus[i+j]), bus[i*2+1], bus[i+:4], bus[i-:4]
- Add epoch guard to prevent redundant rebind evaluations
- Reject mixed local/design-state compound index expressions

+++ include/lyra/common/bit_target_mapping.hpp
@@ -10,7 +10,7 @@ namespace lyra::runtime {
 // Computed once at lowering time. Used identically by codegen and runtime.
 struct BitTargetMapping {
   int32_t index_base = 0;
-  int8_t index_step = 1;
+  int32_t index_step = 1;
   uint32_t total_bits = 0;
 };
